+++ src/lib.rs
@@ -20,8 +20,8 @@ extern crate protobuf;
 extern crate fnv;
 extern crate hyper;
 
-pub mod http;
 pub mod proto;
+pub mod errors;
 pub mod encoder;
 
 mod metrics;
@@ -29,12 +29,11 @@ mod desc;
 // TODO: remove dead_code later.
 #[allow(dead_code)]
 mod value;
-mod errors;
 #[allow(dead_code)]
 mod counter;
 #[allow(dead_code)]
 mod registry;
 
-pub use self::registry::Registry;
+pub use self::registry::{Registry, scrap_registry};
 pub use self::metrics::{Collector, Opts};
 pub use self::counter::Counter;
